refactor(hooks): moved courses_new.go from web to web/hooks
We now do course creation via the github app (tiggered via the app
install event). Therefore, we move the logic from web.createCourse
to hooks.CreateCourse (exporting it). This allows to retain the
tests that rely on the CreateCourse rpc while refactoring.

That is, the handleInstallationCreated now reuses the same function
CreateCourse as for the course creation from the CreateCourse rpc,
which were previously unexported in web.createCourse.

This also adds logging when the course was successfully created.
