+++ web/hooks/installation.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/google/go-github/v62/github"
 	"github.com/quickfeed/quickfeed/qf"
-	"github.com/quickfeed/quickfeed/scm"
 )
 
 func (wh GitHubWebHook) handleInstallationCreated(event *github.InstallationEvent) {
@@ -41,34 +40,12 @@ func (wh GitHubWebHook) handleInstallationCreated(event *github.InstallationEven
 		wh.logger.Errorf("Could not create SCM client for course %s: %v", orgName, err)
 		return
 	}
-	repos, err := sc.CreateCourse(ctx, &scm.CourseOptions{
-		CourseCreator:  courseCreator.Login,
-		OrganizationID: orgID,
-	})
+	c, err := CreateCourse(ctx, wh.db, sc, course, courseCreator)
 	if err != nil {
 		wh.logger.Errorf("Could not create course %s: %v", orgName, err)
 		return
 	}
-	for _, repo := range repos {
-		dbRepo := qf.Repository{
-			ScmOrganizationID: orgID,
-			ScmRepositoryID:   repo.ID,
-			HTMLURL:           repo.HTMLURL,
-			RepoType:          qf.RepoType(repo.Repo),
-		}
-		if dbRepo.IsUserRepo() {
-			dbRepo.UserID = courseCreator.ID
-		}
-		if err := wh.db.CreateRepository(&dbRepo); err != nil {
-			wh.logger.Errorf("Could not create database repository %s: %v", repo.Repo, err)
-			return
-		}
-	}
-
-	if err := wh.db.CreateCourse(courseCreator.ID, course); err != nil {
-		wh.logger.Errorf("Could not create database record for course %s: %v", orgName, err)
-		return
-	}
+	wh.logger.Infof("Successfully created course %v", c)
 
 	if err := wh.tm.Add(courseCreator.ID); err != nil {
 		wh.logger.Errorf("Could not add user %s for token refresh: %v", courseCreator.Login, err)
